Simplify headers handling in the requests
This improvement creates a new attr to be instantiated during the initializer
and after the instantiation it could not be updated directly, just copies of it.

The general idea was to remove some helper methods `clone_options` and `merge_options`
that only had a purpouse because we have been handling the headers inside a hash (options hash).
This extra complexity makes the requirement to have all these internal structures
in order to be possible to add/remove and modify the items inside of it.

After this we should only update a clone when it's required like was did in the lines 30, 41, 72.