Accept request options for all http methods
Before this PR, this was only done for http_post as a workaround to
allow disabling the automatic JSON conversion of "body" to support CSV,
NDJSON, and possible other future payload types.

This must be done for at least http_put at least, since there are
update_* methods to go along with the add_* methods for the various
payload types.

I added the option to all of the http methods to prevent any future
surprises.